Added ChartView component within GraphView.
